refactor: extract room logic into custom hook
- Move room-related logic from RoomDetailPage into useRoomId custom hook
- Create new hooks directory under rooms/[roomId]
- Improve code organization and reusability
- No functional changes
